% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/primetime_wrapper.R
\name{primetime}
\alias{primetime}
\title{Estimation of the PRIMEtime-CE Obesity model}
\usage{
primetime(time.horizon = NULL, age.range = c(20, 100),
  cost.of.treatment = NULL, bmi.target.min = NULL, bmi.target.max = NULL,
  mean.wt.loss.yr1 = NULL, se.diff = NULL, time.to.trt.effect = NULL,
  time.to.weight.regain = NULL, some.wt.loss.maintained = FALSE,
  mean.wt.loss.maintained = NULL, discount.rate.health = NULL,
  discount.rate.cost = NULL, deterministic = TRUE,
  bmi.directEffect.onMortality = TRUE, bmi.directEffect.onQoL = FALSE,
  model.incidence.trends = TRUE, whole.uk.population = TRUE,
  population.characteristics = NULL)
}
\arguments{
\item{time.horizon}{A positive integer. Outcomes are estimated up to age 100 years.}

\item{age.range}{A two element numeric vector giving the lowest and highest ages over which to estimate outcomes.}

\item{cost.of.treatment}{A numeric vector with length equal to the number of active interventions giving intervention costs in year 1.}

\item{bmi.target.min}{A positive integer giving the lowest BMI group at whom the intervention(s) is targeted.}

\item{bmi.target.max}{A positive integer giving the highest BMI group at whom the intervention(s) is targeted. A value of Inf means no upper bound.}

\item{mean.wt.loss.yr1}{A numeric vector with length equal to the number of active interventions to assess. Values are mean weight loss in kilograms 1-year following intervention.}

\item{time.to.trt.effect}{A non-negative integer giving the delay between initiation of treatment and health effects.}

\item{time.to.weight.regain}{A positive integer giving the years till weight returns to baseline, or in the case of some long-term residual weight difference, the time till the weight difference stabilises.}

\item{some.wt.loss.maintained}{A logical value indicating whether there is some residual weight difference between treatment and the control groups after weight regain.}

\item{mean.wt.loss.maintained}{A numeric vector with length equal to the number of active interventions to assess. Values are mean weight loss in kilograms.}
}
\value{
Mean per-person health and economic outcomes in groups defined by sex and age (in 5-year bands) for each year of analysis, for the control and treatment groups, and differences between treatment and control groups. The output can be accessed by other functions to produce standardised outputs (e.g. summary tables), or can be used directly by the user for bespoke analyses.
}
\description{
\code{primetime} generates estimates of disease incidence, life-years, quality-adjusted life-years, and health and social care costs over the specified time-horizon up to age 100 years by sex and age group (in 5-year bands).
}
\examples{
primetime(time.horizon = 20, age.range = c(20, 30), bmi.target.min = 30, bmi.target.max = 50, mean.wt.loss.yr1 = 4, time.to.trt.effect = 1, time.to.weight.regain = 5, some.wt.loss.maintained = FALSE, mean.wt.loss.maintained = 0, bmi.min.risk = 21, discount.rate.health = 3.5, discount.rate.cost = 3.5, deterministic = TRUE)
}
