% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/primetimeCE.R
\docType{package}
\name{primetimeCE}
\alias{primetimeCE}
\alias{primetimeCE-package}
\title{primetimeCE: An implementation of the PRIMEtime-CE Obesity model}
\description{
The PRIMEtime package provides three categories of important functions:
primetime...
}
\section{PRIMEtimeCE functions}{

The PRIMEtimeCE functions ...
}

