% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_summaries.R
\name{summary_table}
\alias{summary_table}
\alias{summary_psa}
\title{Summarise results from the PRIMEtime model}
\usage{
summary_table(model.results = NULL, comparator = "control",
  active.intervention = "trt1", extended.summary = FALSE, timeH = NULL,
  costs.to.include = "disease-related nhs costs",
  present.only.overall.results = FALSE, nicely.presented.results = TRUE)

summary_psa(model = NULL, psa.out = NULL)
}
\arguments{
\item{model.results}{The output generated by \code{primetime}.}

\item{comparator}{The comparator intervention, either "control" or "trt" with a numeric end. The interventions selected must have been estimated using \code{primetime}.}

\item{active.intervention}{The new intervention, "trt" with a numeric end.  The interventions selected must have been estimated using \code{primetime}.}

\item{extended.summary}{If FALSE (the default), \code{summary_table} presents only life-years, QALYs, total costs, and the ICER. If TRUE then it additionally presents each type of cost separately and disease incidence if the modelled conditions.}

\item{timeH}{The time horizon over which to present outcomes. By default it selects the time horizon specified in \code{primetime}. However any non-negative number less than this can also be selected.}

\item{costs.to.include}{Which cost types to include and to make up total costs. Following guidelines for health technology appraisal from the National Institute for Health and Care Excellence, the default is to inlcude only NHS relating to the modelled diseases, i.e. "disease-related nhs costs". Alternatively, total costs can additionally include unrelated nhs costs - "unrelated nhs costs" -, disease-related nhs and formal care costs - "disease-related nhs and social care" -, or all nhs and formal care, related and unrelated to the modelled conditions - "all nhs and social care".}

\item{present.only.overall.results}{Whether to present results averaged over all age and sex groups or by sex and age in 15-year bands (default).}

\item{nicely.presented.results}{Whether to present results nicely formatted for presentation (default) or not. If the results are to be used for further calculations or to create graphs, then select FALSE.}

\item{psa.out}{The results returned from running \code{primetime_psa}. This analysis should match that used to generate \code{model.out}.}
}
\value{
Presents mean life-years, quality-adjusted life-years, total costs, and costs by type for two interventions, and the differences in these outcomes between interventions, as well as the incremental cost-effectiveness ratio (ICER). For the PSA summaries, standard errors around differences are reported. 95\% confidence intervals are formed assuming the mean differences follow normal distributions.
}
\description{
\code{summary_table} Summarises the results from the PRIMEtime model. \code{summary_psa} summarises the results adding standard errors and 95\% confidence intervals to estimated differences.
}
\examples{
summary_table(model.results = primetime.results, comparator = "trt1", active.intervention = "trt2",
nicely.presented.results = TRUE)
}
